✨ improve(patch): improve @roots/bud-esbuild exports map
