+++ examples/esbuild/bud.config.js
@@ -1,3 +1,3 @@
-export default async bud => {
-  bud.html().entry({app: 'app.js'})
-}
+import {bud} from '@roots/bud'
+
+bud.html().entry({app: 'app.js'})
